match_results type changed from vector to set to speed up matching speed (to avoid repeated elements in the resut). a new variable added to count the maximum size of treeIff array. its important to know this because it has a great impact on the matching speed. specially when we lower the THRESHOLD_DEPTH, we increase the size of treeIff array. (The largest treeIff array that I found when I ran the MST wkld was 3290)
Also I used to store the entire filter in a bitset of size 192. Now I use a class called bv192 which is
composed of 3 unsigned long variables to store 192 bits. Although the memory usage is the same as before but
bv192 has a faster subset_check method. with normal bitset, you have to make a temporary copy of it to do the
subset check which can slow down the matching process.