Changes has been made to following files:
	modified:   back_end.cc
	modified:   gpu.cu
	modified:   gpu.hh
	modified:   parameters.hh

In this version, we changed the way we output interface-pairs and eliminated a call
to GPU (async_set_zero) for each batch.
Kernels are optimized to use only registers instead of arrays (on local memory).
This complicated the code but improved the performance.
Becasue of these modifications, the size of the batch right now has to be smaller than 256.